+++ review/src/main/java/com/homeaid/service/ReviewService.java
@@ -2,6 +2,7 @@ package com.homeaid.service;
 
 
 import com.homeaid.domain.Review;
+import com.homeaid.domain.enumerate.UserRole;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
@@ -17,4 +18,6 @@ public interface ReviewService {
   Page<Review> getReviewOfWriter(Long writerId, Pageable pageable);
 
   Page<Review> getReviewOfTarget(Long targetId, Pageable pageable);
+
+  Long getReviewTargetInfo(Long reservationId, Long userId, UserRole userRole);
 }
